Add files via upload
After opening up python, I created a new file like the instructions said. Then I saved the untitled file as main. I then ran the module and saved the file that popped up as main. Then i submitted it. I wasn't sure what to name the second file. Also I'm not sure if I was suppose to upload it to this repository.